refactor: 인증 로직과 멤버 로직의 분리(서비스계층)
인증 로직과 멤버 로직의 분리
