+++ src/main/java/gift/controller/auth/AuthController.java
@@ -3,7 +3,7 @@ package gift.controller.auth;
 import gift.dto.AuthResponse;
 import gift.dto.LoginRequest;
 import gift.dto.RegisterRequest;
-import gift.service.MemberService;
+import gift.service.auth.AuthService;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,21 +15,21 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/members")
 public class AuthController {
 
-    private final MemberService memberService;
+    private final AuthService authService;
 
-    public AuthController(MemberService memberService) {
-        this.memberService = memberService;
+    public AuthController(AuthService authService) {
+        this.authService = authService;
     }
 
     @PostMapping("/register")
     public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest registerRequest) {
-        var auth = memberService.register(registerRequest);
+        var auth = authService.register(registerRequest);
         return ResponseEntity.ok(auth);
     }
 
     @PostMapping("/login")
     public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest loginRequest) {
-        var auth = memberService.login(loginRequest);
+        var auth = authService.login(loginRequest);
         return ResponseEntity.ok(auth);
     }
 }
