+++ src/test/java/gift/controller/ProductControllerTest.java
@@ -3,7 +3,7 @@ package gift.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import gift.dto.LoginRequest;
 import gift.dto.ProductRequest;
-import gift.service.MemberService;
+import gift.service.auth.AuthService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -26,15 +26,15 @@ class ProductControllerTest {
     @Autowired
     private ObjectMapper objectMapper;
     @Autowired
-    private MemberService memberService;
+    private AuthService authService;
     private String managerToken;
     private String memberToken;
 
     @BeforeEach
     @DisplayName("관리자, 이용자의 토큰 값 세팅하기")
     void setAccessToken() throws Exception {
-        managerToken = memberService.login(new LoginRequest("admin@naver.com", "password")).token();
-        memberToken = memberService.login(new LoginRequest("member@naver.com", "password")).token();
+        managerToken = authService.login(new LoginRequest("admin@naver.com", "password")).token();
+        memberToken = authService.login(new LoginRequest("member@naver.com", "password")).token();
     }
 
     @Test
