+++ src/test/java/gift/service/MemberServiceTest.java
@@ -7,6 +7,7 @@ import gift.exception.DuplicatedEmailException;
 import gift.exception.InvalidLoginInfoException;
 import gift.exception.NotFoundElementException;
 import gift.model.MemberRole;
+import gift.service.auth.AuthService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -17,6 +18,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 class MemberServiceTest {
 
     @Autowired
+    private AuthService authService;
+    @Autowired
     private MemberService memberService;
     @Autowired
     private AuthInterceptor authInterceptor;
@@ -25,7 +28,7 @@ class MemberServiceTest {
     @DisplayName("회원가입 시도하기 - 성공")
     void registerSuccess() {
         var registerRequest = new RegisterRequest("테스트", "test@naver.com", "testPassword", "MEMBER");
-        var auth = memberService.register(registerRequest);
+        var auth = authService.register(registerRequest);
         var id = authInterceptor.getMemberIdWithToken(auth.token());
         var role = authInterceptor.getMemberRoleWithToken(auth.token());
 
@@ -38,10 +41,10 @@ class MemberServiceTest {
     @DisplayName("중복된 이메일로 회원가입 시도하기 - 실패")
     void registerFailWithDuplicatedEmail() {
         var registerRequest = new RegisterRequest("테스트", "test@naver.com", "testPassword", "MEMBER");
-        var auth = memberService.register(registerRequest);
+        var auth = authService.register(registerRequest);
         var id = authInterceptor.getMemberIdWithToken(auth.token());
 
-        Assertions.assertThatThrownBy(() -> memberService.register(registerRequest)).isInstanceOf(DuplicatedEmailException.class);
+        Assertions.assertThatThrownBy(() -> authService.register(registerRequest)).isInstanceOf(DuplicatedEmailException.class);
 
         memberService.deleteMember(id);
     }
@@ -50,10 +53,10 @@ class MemberServiceTest {
     @DisplayName("로그인 실행하기 - 성공")
     void loginSuccess() {
         var registerRequest = new RegisterRequest("테스트", "test@naver.com", "testPassword", "MEMBER");
-        memberService.register(registerRequest);
+        authService.register(registerRequest);
 
         var loginRequest = new LoginRequest("test@naver.com", "testPassword");
-        var auth = memberService.login(loginRequest);
+        var auth = authService.login(loginRequest);
         var id = authInterceptor.getMemberIdWithToken(auth.token());
         var role = authInterceptor.getMemberRoleWithToken(auth.token());
 
@@ -66,10 +69,10 @@ class MemberServiceTest {
     @DisplayName("로그인 실행하기 - 실패")
     void loginFailWithWrongPassword() {
         var registerRequest = new RegisterRequest("테스트", "test@naver.com", "testPasswords", "MEMBER");
-        var auth = memberService.register(registerRequest);
+        var auth = authService.register(registerRequest);
         var loginRequest = new LoginRequest("test@naver.com", "testPassword");
 
-        Assertions.assertThatThrownBy(() -> memberService.login(loginRequest)).isInstanceOf(InvalidLoginInfoException.class);
+        Assertions.assertThatThrownBy(() -> authService.login(loginRequest)).isInstanceOf(InvalidLoginInfoException.class);
 
         var id = authInterceptor.getMemberIdWithToken(auth.token());
         memberService.deleteMember(id);
@@ -80,14 +83,14 @@ class MemberServiceTest {
     void deleteMemberSuccess() {
         var registerRequest = new RegisterRequest("테스트", "test@naver.com", "testPassword", "MEMBER");
         var loginRequest = new LoginRequest("test@naver.com", "testPassword");
-        var auth = memberService.register(registerRequest);
+        var auth = authService.register(registerRequest);
         var id = authInterceptor.getMemberIdWithToken(auth.token());
-        var loginAuth = memberService.login(loginRequest);
+        var loginAuth = authService.login(loginRequest);
 
         Assertions.assertThat(auth.token()).isEqualTo(loginAuth.token());
 
         memberService.deleteMember(id);
 
-        Assertions.assertThatThrownBy(() -> memberService.login(loginRequest)).isInstanceOf(NotFoundElementException.class);
+        Assertions.assertThatThrownBy(() -> authService.login(loginRequest)).isInstanceOf(NotFoundElementException.class);
     }
 }
